conversationalAI/requirements.txt
packaging==21.0
geocoder==1.38.1
transformers==4.11.3
inflect==5.3.0
playsound==1.3.0
SoundFile==0.10.3.post1
torch==1.9.1
tensorflow==2.6.0
SpeechRecognition==3.8.1
openai==0.10.5
scipy==1.7.1
numpy==1.20.3
pandas_stubs==1.2.0.19
Unidecode==1.3.2
pyworld==0.3.0
absl_py==0.14.1
phonemizer==2.2.2
matplotlib==3.4.3
tqdm==4.62.3
Cython==0.29.24
librosa==0.8.1
absl==0.0
pandas==1.3.5
Pillow==9.0.0
PySimpleGUI==4.56.0
PyYAML==6.0
tensorboardX==2.4.1
umap==0.1.1
